Create Command Handle
