+++ commands/hey.js
@@ -1,0 +1,5 @@
+const Discord = require("discord.js");
+
+module.exports.run = async (client, message,args ) => {
+    message.channel.send("hey! How can i help you ");
+}
